backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
djangorestframework-simplejwt>=5.2.2
APScheduler>=3.10.4
